refactor: move pot events
