+++ lib/app/modules/socket/data/models/pot_events/chat_v1_event.dart
