+++ lib/app/modules/socket/data/models/pot_events/user_in_v1_event.dart
